better modes select in webhooks/crons/etc
